Snooping
Changed all the scrolling stuff to rely on textOffset, instead of moving an oversized & clipped box. Also tried using the getters and setters of TextField, to make sure they were sending events, in case that had anything to do with it. The new approach works, but identically, with the same inability to select text after line 32.