Added G as alt gate symbol
Felt weird to label it 'G' internally but not use 'G' as the command, but kept X as an acceptable variation. Maybe 'X' can become some sort of "autogate", that decides whether to be a gate or trigger by context somehow, if there's some definition that makes sense for that.